+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Check/create directory before saving ([#196](https://github.com/getsentry/sentry-unity/pull/196))
 - Embedded link.xml in assembly ([#194](https://github.com/getsentry/sentry-unity/pull/194))
 - Exclude SentryOptions.json from release package ([#195](https://github.com/getsentry/sentry-unity/pull/195))
+- Transition scene to test scene loading events (breadcrumbs) ([#185](https://github.com/getsentry/sentry-unity/pull/185))
 
 ## 0.0.14