+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Fixes
 
+- empty override fields in config window ([#198](https://github.com/getsentry/sentry-unity/pull/198))
 - Check/create directory before saving ([#196](https://github.com/getsentry/sentry-unity/pull/196))
 - Exclude SentryOptions.json from release package ([#195](https://github.com/getsentry/sentry-unity/pull/195))